add tests for cookies
